Add confirmation alert

+++ frontend/app/assets/javascripts/spree/frontend/checkout.js
@@ -61,15 +61,17 @@ Spree.ready(function () {
 })
 
 $('.js-remove-credit-card').click(function() {
-  return $.ajax({
-    async: false,
-    method: 'DELETE',
-    url: Spree.routes.api_v2_storefront_destroy_credit_card(this.dataset.id),
-    dataType: 'json',
-    headers: {
-      'Authorization': 'Bearer ' + SpreeAPI.oauthToken
-    }
-  }).done(function() {
-    location.reload();
-  })
+  if (confirm(Spree.translations.credit_card_remove_confirmation)) {
+    return $.ajax({
+      async: false,
+      method: 'DELETE',
+      url: Spree.routes.api_v2_storefront_destroy_credit_card(this.dataset.id),
+      dataType: 'json',
+      headers: {
+        'Authorization': 'Bearer ' + SpreeAPI.oauthToken
+      }
+    }).done(function() {
+      location.reload();
+    })
+  }
 })
